// src/services/client/models/PublicQuestionSerializerDetail.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { FieldTypeEnum } from './FieldTypeEnum';

export type PublicQuestionSerializerDetail = {
    key: string;
    field_type: FieldTypeEnum;
    meta: Record<string, any>;
    required: boolean;
};